APOD client: Fetches APOD for chosen date
From the menu/title, press [<] and [>] to change the date.
Press [=] to choose "the current date".  (The request to the
server will not specify the date, so always fetch the latest.
Helpful for when your timezone isn't caught up with the current
day in the Eastern US timezone; e.g., at 9:00pm Pacific the
"next day's" APOD will be available.)

Also, running out of space; removed the colorbars used
by non-ColorView9 modes, to save a little space.