Attempting to get Eastern Timezone via APETIME
Expanding APETIME functionality in FujiNet-PlatformIO
to allow specification of the timezone we want, when calling
a new variation of GETTIME.

Note - Not 100% working; hangs when it writes the request to
set the time zone.  FujiNet seems to get it fine, however, and
hitting [Break] on the Atari proceeds.

Motivation - APOD photos change at 12:00am (00:00) each day,
Eastern time.  (So for me, in Pacific time zone, FujiNet doesn't
know there's "a new day" for 3 hours.  Folks east of ET can
have a localtime date that's in the future, compared to ET.)

See https://github.com/FujiNetWIFI/fujinet-platformio/issues/449