ISS Tracker Atari: Use #FujiNet JSON parsing: Tracking
Note - The tracking functionality does not work right now,
because the timestamp we get during the initial positioning
comes back in E-notation, which the program is unable to parse
(and it would lose about 10K seconds of resolution, or almost 3hrs,
even if I was able to parse it).

See https://github.com/FujiNetWIFI/fujinet-platformio/issues/500

Once that issue is addressed somehow, it should work like it did
previously (with the `faux_json` JSON parsing code that was running
on the Atari).

One benefit here is we can now query 10 tracking positons at
a time, and only do two HTTP requests to get 20 positions
(rather than 2 positions at a time, requiring 10 HTTP requests).