Use usual Apple II Ethernet cards as alternative to FujiNet.
IP65 offers the url_download() API that fits in nicely as there's no TLS involved. JSON parsing is replaced with - of course unflexible - scanf() usage.

The two IP65 libraries are renamed as they need to be linked in a certain order and the given Makefile sets the order implicitly by name.