Tweaks to apple-tool scripts
Don't echo java path while running by silencing the which command.

Make add-file.sh take an arg for autostarting, it's job was originally
designed to just add files to disk, so default is not to make them
autostart.

Fix makefile scripts that used this feature to add a param for autostarting.